+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
  - Add `windows` release
  - Add `darwin-386` release
+ - Start counting for `lines` in `Stdout` and `Stderr` at `1` instead of `0`
 
 # v0.2.1